+++ react/components/Menu/MenuLink.js
@@ -8,7 +8,7 @@ const MenuLink = ({ path, name, location }) => {
     <Link
       to={path}
       className={classnames(
-        'f6 no-underline db hover-near-black pv5 mv3 pl5 bl bw2',
+        'f6 no-underline db hover-near-black pv5 mv3 pl5 bl bw2 nowrap',
         { 'mid-gray b--transparent': location.pathname.indexOf(path) === -1 },
         { 'near-black b b--blue': location.pathname.indexOf(path) !== -1 },
       )}