+++ react/components/Menu/MenuLinksList.js
@@ -1,12 +1,9 @@
 import React from 'react'
 import { injectIntl, intlShape } from 'react-intl'
+import { ExtensionPoint } from 'render'
 import MenuLink from './MenuLink'
 
 const links = [
-  {
-    id: 'pages.orders',
-    path: '/my-orders',
-  },
   {
     id: 'pages.addresses',
     path: '/addresses',
@@ -24,15 +21,25 @@ const links = [
 const MenuLinksList = ({ intl }) => {
   return (
     <nav>
-      {links.map(link => (
-        <MenuLink
-          path={link.path}
-          name={intl.formatMessage({
-            id: link.id,
-          })}
-          key={link.id}
-        />
+      <ExtensionPoint
+        id="menu-links-before"
+        render={links =>
+          links.map(({ name, path }) => (
+            <MenuLink path={path} name={name} key={name} />
+          ))
+        }
+      />
+      {links.map(({ path, id }) => (
+        <MenuLink path={path} name={intl.formatMessage({ id })} key={id} />
       ))}
+      <ExtensionPoint
+        id="menu-links-after"
+        render={links =>
+          links.map(({ name, path }) => (
+            <MenuLink path={path} name={name} key={name} />
+          ))
+        }
+      />
     </nav>
   )
 }