+++ react/components/Profile/ProfileBox.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { injectIntl, intlShape } from 'react-intl'
-import { ProfileRules, ProfileSummary } from '@vtex/profile-form'
+import { ExtensionPoint } from 'render'
+import { ProfileRules, ProfileSummary } from 'vtex.profile-form'
 import ContentBox from '../shared/ContentBox'
 import DataEntry from '../shared/DataEntry'
 
@@ -15,10 +16,7 @@ const ProfileBox = ({ profile, onEditClick, intl }) => {
       lowerButton={intl.formatMessage({ id: 'commons.edit' })}
       onLowerButtonClick={onEditClick}
     >
-      <ProfileRules
-        country={'BRA'}
-        fetch={country => import('@vtex/profile-form/lib/rules/' + country)}
-      >
+      <ProfileRules country={'BRA'} shouldUseIOFetching>
         <ProfileSummary profile={profile}>
           {({
             personalData: {
@@ -36,44 +34,39 @@ const ProfileBox = ({ profile, onEditClick, intl }) => {
               <div>
                 <div className="flex-ns flex-wrap">
                   <div className="mb8 flex-auto">
-                    <DataEntry
-                      label={firstName.label}
-                      content={firstName.value}
-                    />
+                    <DataEntry label={firstName.label}>
+                      {firstName.value}
+                    </DataEntry>
                   </div>
                   <div className="mb8 flex-auto">
-                    <DataEntry
-                      label={lastName.label}
-                      content={lastName.value}
-                    />
+                    <DataEntry label={lastName.label}>
+                      {lastName.value}
+                    </DataEntry>
                   </div>
                 </div>
                 <div className="mb8">
-                  <DataEntry label={email.label} content={email.value} />
+                  <DataEntry label={email.label}>{email.value}</DataEntry>
                 </div>
                 <div className="flex-ns flex-wrap">
                   <div className="mb8 flex-auto">
-                    <DataEntry
-                      label={document.label}
-                      content={document.value}
-                    />
+                    <DataEntry label={document.label}>
+                      {document.value}
+                    </DataEntry>
                   </div>
                   <div className="mb8 w-50-ns">
-                    <DataEntry label={gender.label} content={gender.value} />
+                    <DataEntry label={gender.label}>{gender.value}</DataEntry>
                   </div>
                 </div>
                 <div className="flex-ns flex-wrap">
                   <div className="mb8 flex-auto">
-                    <DataEntry
-                      label={birthDate.label}
-                      content={birthDate.value}
-                    />
+                    <DataEntry label={birthDate.label}>
+                      {birthDate.value}
+                    </DataEntry>
                   </div>
                   <div className="mb8 w-50-ns">
-                    <DataEntry
-                      label={homePhone.label}
-                      content={homePhone.value}
-                    />
+                    <DataEntry label={homePhone.label}>
+                      {homePhone.value}
+                    </DataEntry>
                   </div>
                 </div>
               </div>
@@ -82,34 +75,29 @@ const ProfileBox = ({ profile, onEditClick, intl }) => {
               ) && (
                 <div>
                   <div className="mb8">
-                    <DataEntry
-                      label={businessData.corporateName.label}
-                      content={businessData.corporateName.value}
-                    />
+                    <DataEntry label={businessData.corporateName.label}>
+                      {businessData.corporateName.value}
+                    </DataEntry>
                   </div>
                   <div className="mb8">
-                    <DataEntry
-                      label={businessData.tradeName.label}
-                      content={businessData.tradeName.value}
-                    />
+                    <DataEntry label={businessData.tradeName.label}>
+                      {businessData.tradeName.value}
+                    </DataEntry>
                   </div>
                   <div className="mb8">
-                    <DataEntry
-                      label={businessData.corporateDocument.label}
-                      content={businessData.corporateDocument.value}
-                    />
+                    <DataEntry label={businessData.corporateDocument.label}>
+                      {businessData.corporateDocument.value}
+                    </DataEntry>
                   </div>
                   <div className="mb8">
-                    <DataEntry
-                      label={businessData.businessPhone.label}
-                      content={businessData.businessPhone.value}
-                    />
+                    <DataEntry label={businessData.businessPhone.label}>
+                      {businessData.businessPhone.value}
+                    </DataEntry>
                   </div>
                   <div className="mb8">
-                    <DataEntry
-                      label={businessData.stateRegistration.label}
-                      content={businessData.stateRegistration.value}
-                    />
+                    <DataEntry label={businessData.stateRegistration.label}>
+                      {businessData.stateRegistration.value}
+                    </DataEntry>
                   </div>
                 </div>
               )}
@@ -117,6 +105,20 @@ const ProfileBox = ({ profile, onEditClick, intl }) => {
           )}
         </ProfileSummary>
       </ProfileRules>
+      <ExtensionPoint
+        id="profile/display"
+        render={fields => (
+          <div className="flex-ns flex-wrap">
+            {fields.map(({ label, value }) => (
+              <div className="mb8 w-50-ns">
+                <DataEntry key={label} label={label}>
+                  {value}
+                </DataEntry>
+              </div>
+            ))}
+          </div>
+        )}
+      />
     </ContentBox>
   )
 }