+++ react/components/Profile/ProfileFormBox.js
@@ -4,7 +4,7 @@ import { intlShape, injectIntl } from 'react-intl'
 import { graphql } from 'react-apollo'
 import { compose } from 'recompose'
 import { Button } from 'vtex.styleguide'
-import { ProfileRules, ProfileContainer } from '@vtex/profile-form'
+import { ProfileRules, ProfileContainer } from 'vtex.profile-form'
 import ContentBox from '../shared/ContentBox'
 import UpdateProfile from '../../graphql/updateProfile.gql'
 
@@ -14,12 +14,18 @@ class ProfileFormBox extends Component {
     this.state = {
       isLoading: false,
     }
+    this.extension = React.createRef()
   }
 
   handleSubmit = async ({ valid, profile: profileInput }) => {
     const { updateProfile, onDataSave, onError } = this.props
     const { email, ...profile } = profileInput
     if (!valid || this.state.isLoading) return
+
+    if (this.extension.current && this.extension.current.submit) {
+      const extensionValid = this.extension.current.submit()
+      if (!extensionValid) return
+    }
 
     try {
       this.setState({ isLoading: true })
@@ -39,10 +45,7 @@ class ProfileFormBox extends Component {
 
     return (
       <ContentBox shouldAllowGrowing maxWidthStep={6}>
-        <ProfileRules
-          country={'BRA'}
-          fetch={country => import('@vtex/profile-form/lib/rules/' + country)}
-        >
+        <ProfileRules country={'BRA'} shouldUseIOFetching>
           <ProfileContainer
             defaultProfile={profile}
             onSubmit={this.handleSubmit}