+++ react/pages/payment-create/PaymentCreateHeader.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import BaseHeader from '../shared/BaseHeader'
+
+const PaymentCreateHeader = () => {
+  return (
+    <BaseHeader
+      titleId={'pages.paymentCreate'}
+      backButton={{ id: 'pages.payments', path: '/payments' }}
+      shouldAlwaysShowBackButton
+    />
+  )
+}
+
+export default PaymentCreateHeader