+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.12.0] - 2018-08-20
+
+### Added
+
+- Extension points
+- Payment list (mocked)
+
 ## [0.11.0] - 2018-08-15
 
 ## Added